requirements.txt
argparse
munch
requests
tabulate
tqdm
xmltodict
pytimeparse # for parsing time intervals
pyyaml
